+++ src/main.rs
@@ -13,6 +13,7 @@ use std::io::Cursor;
 use std::io::Read;
 use std::path::PathBuf;
 use std::process::ExitCode;
+use std::sync::{Arc, atomic::AtomicBool};
 use tracing_subscriber::{
     EnvFilter, Layer, fmt::format::FmtSpan, layer::SubscriberExt, util::SubscriberInitExt,
 };
@@ -139,10 +140,12 @@ fn main() -> ExitCode {
         font: cli.options.font.unwrap_or(String::from("Noto Sans CJK JP")),
     };
 
+    let paused = Arc::new(AtomicBool::new(false));
+
     #[cfg(target_os = "linux")]
     {
         if config.show_tray_icon {
-            crate::tray::spawn_tray_icon();
+            crate::tray::spawn_tray_icon(Arc::clone(&paused));
         }
 
         if let Some(text) = &cli.modes.text {
@@ -166,7 +169,7 @@ fn main() -> ExitCode {
                 return ExitCode::FAILURE;
             }
         } else if cli.modes.watch {
-            if let Err(e) = popup_dictionary::watch(config) {
+            if let Err(e) = popup_dictionary::watch(config, Arc::clone(&paused)) {
                 tracing::error!("Failed while running watch mode due to error: {e}");
                 return ExitCode::FAILURE;
             }
@@ -189,9 +192,9 @@ fn main() -> ExitCode {
             tracing::info!("No mode specified. Defaulting to watch mode with tray icon.");
             // Default to watch mode with tray icon if no mode set
             if !config.show_tray_icon {
-                crate::tray::spawn_tray_icon();
+                crate::tray::spawn_tray_icon(Arc::clone(&paused));
             }
-            if let Err(e) = popup_dictionary::watch(config) {
+            if let Err(e) = popup_dictionary::watch(config, Arc::clone(&paused)) {
                 tracing::error!("Failed while running watch mode due to error: {e}");
                 return ExitCode::FAILURE;
             }
@@ -200,7 +203,7 @@ fn main() -> ExitCode {
     #[cfg(target_os = "windows")]
     {
         if config.show_tray_icon {
-            crate::tray::spawn_tray_icon();
+            crate::tray::spawn_tray_icon(Arc::clone(&paused));
         }
 
         if let Some(text) = &cli.modes.text {
@@ -214,7 +217,7 @@ fn main() -> ExitCode {
                 return ExitCode::FAILURE;
             }
         } else if cli.modes.watch {
-            if let Err(e) = popup_dictionary::watch(config) {
+            if let Err(e) = popup_dictionary::watch(config, Arc::clone(&paused)) {
                 tracing::error!("Failed while running watch mode due to error: {e}");
                 return ExitCode::FAILURE;
             }
@@ -237,9 +240,9 @@ fn main() -> ExitCode {
             tracing::info!("No mode specified. Defaulting to watch mode with tray icon.");
             // Default to watch mode with tray icon if no mode set
             if !config.show_tray_icon {
-                crate::tray::spawn_tray_icon();
+                crate::tray::spawn_tray_icon(Arc::clone(&paused));
             }
-            if let Err(e) = popup_dictionary::watch(config) {
+            if let Err(e) = popup_dictionary::watch(config, Arc::clone(&paused)) {
                 tracing::error!("Failed while running watch mode due to error: {e}");
                 return ExitCode::FAILURE;
             }